♻️ Refactor: ChatMessageGroup 컴포넌트로 로직 분리 및 타입 추가 (#86)
