+++ src/pages/chat/sections/chatMessageList/ChatMessageList.tsx
@@ -1,13 +1,8 @@
-import clsx from 'clsx';
 import { useEffect } from 'react';
 
-import { ChatTimeStyles } from '../../chat.styles';
 import type { ChatTypes, GroupedChatListTypes } from '../../chat.types';
 import { CHAT_EXAMPLES } from '../../chatSampleData';
-import ChatMessageItem from './ChatMessageItem';
-
-// *memo - 로그인 시 전역 상태로 저장되는 userId 값이 존재한다고 가정
-const USER_ID = 'idol-01';
+import ChatMessageGroup from './ChatMessageGroup';
 
 function ChatMessageList() {
   const toKstDate = (iso: string | Date): Date =>
@@ -73,26 +68,13 @@ function ChatMessageList() {
             {dKey}
           </div>
 
-          {Object.entries(dValue).map(([tKey, tValue]) => {
-            const isLastMsgMine = tValue.at(-1)?.sender.id === USER_ID;
-
-            return (
-              <div key={`T-${dKey}-${tKey}`}>
-                {tValue.map(msg => (
-                  <ChatMessageItem key={msg.id} {...msg} />
-                ))}
-
-                {isLastMsgMine}
-                <div
-                  className={clsx(
-                    ChatTimeStyles({ lastMsgMine: isLastMsgMine }),
-                  )}
-                >
-                  {tKey}
-                </div>
-              </div>
-            );
-          })}
+          {Object.entries(dValue).map(([tKey, tValue]) => (
+            <ChatMessageGroup
+              key={`T-${dKey}-${tKey}`}
+              tKey={tKey}
+              tValue={tValue}
+            />
+          ))}
         </div>
       ))}
     </div>
